change conditions

+++ 9-binary_tree_height.c
@@ -8,7 +8,7 @@
  */
 size_t binary_tree_height(const binary_tree_t *tree)
 {
-	size_t left = 0, right = 0;
+	int left = 0, right = 0;
 
 	if (!tree)
 		return (0);
